# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateOrganizationRequest:
    properties:
      active:
        example: true
        type: boolean
      name:
        example: Acme Corp
        type: string
    required:
    - name
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handlers.UpdateOrganizationRequest:
    properties:
      active:
        example: false
        type: boolean
      name:
        example: Acme Corp Updated
        type: string
    type: object
  models.Group:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Administrators
        type: string
      organizations:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - name
    type: object
  models.Organization:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        example: 1
        type: integer
      name:
        example: Acme Corp
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - name
    type: object
  models.User:
    properties:
      active:
        example: true
        type: boolean
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      created_by:
        example: admin@example.com
        type: string
      email:
        example: john@example.com
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      organizations:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    required:
    - email
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@kitamanager.example.com
    name: API Support
  description: REST API for managing Users, Groups, and Organizations with RBAC support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: KitaManager API
  version: "1.0"
paths:
...
  /api/v1/organizations:
    get:
      consumes:
      - application/json
      description: Get a list of all organizations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Organization'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Create a new organization
      parameters:
      - description: Organization data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Organization'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new organization
      tags:
      - organizations
  /api/v1/organizations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an organization
      tags:
      - organizations
    get:
      consumes:
      - application/json
      description: Get a single organization by its ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organization'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get organization by ID
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Update an existing organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: integer
      - description: Organization data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organization'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an organization
      tags:
      - organizations
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"